+++ pwm.h
@@ -10,21 +10,21 @@
 
 #include <stdint.h>
 
-#define PWMIR   (*(volatile unsigned long *) 0xE0014000)
-#define PWMTCR  (*(volatile unsigned long *) 0xE0014004)
-#define PWMTC   (*(volatile unsigned long *) 0xE0014008)
-#define PWMPR   (*(volatile unsigned long *) 0xE001400C)
-#define PWMPC   (*(volatile unsigned long *) 0xE0014010)
-#define PWMMCR  (*(volatile unsigned long *) 0xE0014014)
-#define PWMMR0  (*(volatile unsigned long *) 0xE0014018)
-#define PWMMR1  (*(volatile unsigned long *) 0xE001401C)
-#define PWMMR2  (*(volatile unsigned long *) 0xE0014020)
-#define PWMMR3  (*(volatile unsigned long *) 0xE0014024)
-#define PWMMR4  (*(volatile unsigned long *) 0xE0014040)
-#define PWMMR5  (*(volatile unsigned long *) 0xE0014044)
-#define PWMMR6  (*(volatile unsigned long *) 0xE0014048)
-#define PWMPCR  (*(volatile unsigned long *) 0xE001404C)
-#define PWMLER  (*(volatile unsigned long *) 0xE0014050)
+#define PWMIR   (*(volatile uint32_t *) 0xE0014000)
+#define PWMTCR  (*(volatile uint32_t *) 0xE0014004)
+#define PWMTC   (*(volatile uint32_t *) 0xE0014008)
+#define PWMPR   (*(volatile uint32_t *) 0xE001400C)
+#define PWMPC   (*(volatile uint32_t *) 0xE0014010)
+#define PWMMCR  (*(volatile uint32_t *) 0xE0014014)
+#define PWMMR0  (*(volatile uint32_t *) 0xE0014018)
+#define PWMMR1  (*(volatile uint32_t *) 0xE001401C)
+#define PWMMR2  (*(volatile uint32_t *) 0xE0014020)
+#define PWMMR3  (*(volatile uint32_t *) 0xE0014024)
+#define PWMMR4  (*(volatile uint32_t *) 0xE0014040)
+#define PWMMR5  (*(volatile uint32_t *) 0xE0014044)
+#define PWMMR6  (*(volatile uint32_t *) 0xE0014048)
+#define PWMPCR  (*(volatile uint32_t *) 0xE001404C)
+#define PWMLER  (*(volatile uint32_t *) 0xE0014050)
 
 #define PWM_COUNTER_RANGE       4096
